EES-659 - refactor
